refactor(settings): separate aria runtime ownership
