+++ docs/design-docs/module-boundary-naming-audit.md
@@ -2,8 +2,8 @@
 
 Status: maintained verified audit
 Last verified: 2026-07-28
-Verification base: `e29ecc8ceba0ca4279187929aba7733ee2619d97` plus the Gate 9 SQLite-store working tree
-Verification branch: `refactor/sqlite-download-store`
+Verification base: `9f570f4bf2e4bba75b15b749b9e979567005467e` plus the Gate 9 settings-owner working tree
+Verification branch: `refactor/network-settings-owner`
 
 ## 結論
 
@@ -25,7 +25,7 @@ pwsh ./script/audit-module-boundaries.ps1 `
 | Source root | Files | Physical lines |
 |---|---:|---:|
 | `DownKyi` | 1 | 14 |
-| `DownKyi.Core` | 271 | 18,245 |
+| `DownKyi.Core` | 272 | 18,248 |
 | `src/DownKyi.Domain` | 11 | 681 |
 | `src/DownKyi.Application` | 30 | 1,143 |
 | `src/DownKyi.Infrastructure` | 27 | 3,751 |
@@ -47,7 +47,7 @@ pwsh ./script/audit-module-boundaries.ps1 `
 | resolved | service contracts 依賴 ViewModel | 0 interfaces | completed by Gate 8 |
 | resolved | custom collection contract | 0 custom collection references; standard read-only wrappers | completed by Gate 8 |
 | resolved | naming and folder taxonomy inconsistent | 4 endpoint/role-scoped duplicate groups, 0 generic names, 0 file/type mismatches | completed in Gate 9 naming branch |
-| P2 | oversized owners | 12 production files above 500 physical lines | confirmed, decreasing |
+| P2 | oversized owners | 11 production files above 500 physical lines | confirmed, decreasing |
 | resolved | logging owner too broad | contracts in Application; 268-line provider plus dedicated Infrastructure sink/buffer/retention/export owners | completed by Gate 9 PR #94 |
 | P1 | AI knowledge environment incomplete | required root/docs structure and reproducible audit scripts now exist | resolved |
 
@@ -162,7 +162,7 @@ File/type mismatch baseline 已由 4 項降為 0。Bilibili JSON DTO 與 NFO XML
 
 ## Finding 11: 巨檔 owners
 
-2026-07-28 的實際 boundary audit 有 12 個 production files 超過 500 physical lines；`DownloadPipeline`、`DownloadTaskProjectionStore`、`ViewMyFavoritesViewModel`、`ViewPublicationViewModel`、`SqliteDownloadTaskStore` 與原 715 行 `ApplicationLogProvider` 已從 allowlist 移除。SQLite Store 從 928 行降為 447 行，交易/初始化協調、Domain row mapping、讀取/quarantine 與 SQL writes 已分成具名 owner，既有 schema/migration/resume tests 保持不變。其餘最優先的手寫 owners：
+2026-07-28 的實際 boundary audit 有 11 個 production files 超過 500 physical lines；`DownloadPipeline`、`DownloadTaskProjectionStore`、`ViewMyFavoritesViewModel`、`ViewPublicationViewModel`、`SqliteDownloadTaskStore`、`SettingsManager.Network` 與原 715 行 `ApplicationLogProvider` 已從 allowlist 移除。SQLite Store 從 928 行降為 447 行，交易/初始化協調、Domain row mapping、讀取/quarantine 與 SQL writes 已分成具名 owner，既有 schema/migration/resume tests 保持不變。Settings 的一般 network/downloader/proxy owner 為 319 行，aria RPC/runtime owner 為 355 行；44 個既有 public compatibility methods 的方法級內容相同，且仍讀寫同一個 `ApplicationSettings.Network` schema。其餘最優先的手寫 owners：
 
 - `ViewVideoViewModel.cs` 1,020
 - `ViewMySpaceViewModel.cs` 669
@@ -200,6 +200,7 @@ Logging 風險已依 ADR 收斂並由 PR #94 整合：Application 只保留 cont
 8. UI collection polling 不可擴散。
 9. static/sync HTTP debt 不可擴散。
 10. 已刪除的 custom mutable collection 不得返回；下載清單只能公開標準唯讀 wrapper。
+11. 一般 network 與 aria RPC/runtime settings owners 必須維持分離，且兩個檔案都受 500 行上限約束。
 
 這些測試是過渡 ratchet。每移除一項債務，應同步刪除對應 baseline entry；不得把 baseline 當成永久例外清單。
